requirements.txt
appdirs==1.4.4
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
constantly==15.1.0
cryptography==41.0.4
cssselect==1.2.0
fake-useragent==1.3.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.4
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jieba3k==0.35.1
jmespath==1.0.1
joblib==1.3.2
libretranslatepy==2.1.1
lxml==4.9.3
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.1
packaging==23.2
pandas==2.1.1
parse==1.19.1
parsel==1.8.1
Pillow==10.1.0
Protego==0.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyee==8.2.2
pyOpenSSL==23.2.0
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
queuelib==1.6.2
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-html==0.10.0
Scrapy==2.11.0
service-identity==23.1.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
tinysegmenter==0.3
tldextract==5.0.1
tqdm==4.66.1
translate==3.6.1
Twisted==22.10.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
zipp==3.17.0
zope.interface==6.1
